Global theme colors and rounded percentNumber
