+++ .gitignore
@@ -61,3 +61,5 @@ yarn-error.log
 
 # Temporary files created by Metro to check the health of the file watcher
 .metro-health-check*
+
+apiresponse.json
