+++ App.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import {ThemeProvider} from 'styled-components/native';
 import {SafeAreaView, StatusBar} from 'react-native';
 import {ContentView} from './styles';
-import {lightTheme, darkTheme} from './theme';
+import {lightTheme, darkTheme} from './src/utils/theme';
 import Main from './src/pages/Main/index';
 import {useColorScheme} from 'react-native';
