+++ src/components/CryptCard/index.tsx
@@ -13,20 +13,13 @@ import {
   ThemedText,
 } from './styles';
 
+
 const CryptCard = () => {
   const percentageChange = (percentNumber: number) => {
-    const roundNumber =
-      Math.round((percentNumber + Number.EPSILON) * 100) / 100;
+    const roundNumber = Number(percentNumber.toFixed(2));
     const negative = roundNumber < 0;
-
-    // const roundNumber = useMemo(() => {
-    //   return Math.round((percentNumber + Number.EPSILON) * 100) / 100
-    // }, [percentNumber])
-
-    // const negative = useMemo(() => {
-    //   return roundNumber < 0
-    // }, [roundNumber])
-
+  
+    
     return (
       <PercentChangeView>
         <PercentChangeArrow
@@ -34,21 +27,21 @@ const CryptCard = () => {
           height={16}
           rotation={negative ? 180 : 0}
           fill={negative ? 'red' : 'green'}
-        />
+          />
         <PercentChangeNumber negative={negative}>
           {(negative ? -roundNumber : roundNumber) + '%'}
         </PercentChangeNumber>
       </PercentChangeView>
     );
   };
-
+  
   const memoCryptoPrice = useMemo(
     () =>
       Math.round((crypto.data.market_data.price_usd + Number.EPSILON) * 100) /
       100,
     [crypto.data.market_data.price_usd],
   );
-
+  
   const memoPercentageChange = useMemo(
     () =>
       percentageChange(
@@ -56,7 +49,6 @@ const CryptCard = () => {
       ),
     [crypto.data.market_data.percent_change_usd_last_24_hours],
   );
-
   return (
     <ContainerView key={crypto.data.id}>
       <ColView>
