+++ src/utils/styled.d.ts
@@ -6,6 +6,10 @@ declare module 'styled-components/native' {
 
       primary: string;
       text: string;
-
+      grey: string;
+      green: string;
+      red: string;
+      blue: string;
+      yellow: string;
   }
 }
