+++ src/utils/theme.ts
@@ -1,0 +1,23 @@
+import {DefaultTheme} from 'styled-components';
+
+const darkTheme: DefaultTheme = {
+  primary: 'black',
+  text: 'white',
+  grey: '#f5f5f5',
+  green: 'green',
+  red: 'red',
+  blue: '#385775',
+  yellow: '#f3c30069',
+};
+
+const lightTheme: DefaultTheme = {
+  primary: 'white',
+  text: 'black',
+  grey: '#696969',
+  green: 'green',
+  red: 'red',
+  blue: '#385775',
+  yellow: '#ffc700',
+};
+
+export {darkTheme, lightTheme};
