--- theme.ts
@@ -1,20 +1,0 @@
-import { DefaultTheme } from 'styled-components';
-
-
-const darkTheme: DefaultTheme = {
-
-    primary: 'black',
-    text: 'white',
-
-
-};
-
-const lightTheme: DefaultTheme = {
-
-    primary: 'white',
-    text: 'black',
-
-};
-
-export { darkTheme, lightTheme };
-
